refactor: move the bn254 field definition out of common
With the Mavros prover de-spined, nothing in common's non-test code names
Bn254Field anymore, so the concrete field definition can leave the spine.

Move Bn254Field, its ProofField/FieldHash impls, the WhirConfig/WhirZkConfig
aliases, and the two bn254 hash tests from common into a new field module in
provekit-backend-bn254. The impls reference only common's public surface
(HashConfig, the byte bridges, hash_field_elements), so the bn254 hash
bodies travel cleanly; the orphan rule is satisfied because backend now owns
Bn254Field.

common is now Bn254Field-free; what remains is field-generic
(WhirR1CSScheme<P>, ProvekitProof<P>, WhirR1CSProof, R1csHash). The only
remaining bn254 anchor in the spine is the FieldElement = ark_bn254::Fr
alias (deleted in the final step). r1cs-compiler and gnark repoint to the
backend; gnark gains a backend dependency.

72/72 e2e green; the two relocated hash tests pass in the backend.
