+++ provekit/backend/bn254/src/scheme/prover.rs
@@ -1,10 +1,8 @@
 use {
-    crate::{MavrosProver, NoirElement, NoirProofScheme, NoirWitnessGenerator},
+    crate::{Bn254Field, MavrosProver, NoirElement, NoirProofScheme, NoirWitnessGenerator},
     acir::circuit::Program,
     noirc_abi::Abi,
-    provekit_common::{
-        witness::SplitWitnessBuilders, Bn254Field, HashConfig, WhirR1CSScheme, R1CS,
-    },
+    provekit_common::{witness::SplitWitnessBuilders, HashConfig, WhirR1CSScheme, R1CS},
     serde::{Deserialize, Serialize},
 };
